+++ packages/blocks/core/php/third-party/blocksy-about-me/block.php
@@ -19,6 +19,9 @@ return array_merge(
 				'blockera/elements/icons'   => [
 					'root' => '.ct-icon-container',
 				],
+				'blockera/elements/avatar' => [
+					'root' => 'figure img',
+				],
 			]
 		),
 	]