+++ packages/blocks/core/CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added support for the `Blocksy About Me Block` text inner block by Blockera.
 - Added support for the `Blocksy About Me Block` icons inner block by Blockera.
 - Added support for the `Blocksy About Me Block` icons border color by Blockera.
+- Added support for the `Blocksy About Me Block` avatar inner block by Blockera.
 
 
 ### Automated Tests